+++ mini-typescript-hero/src/configuration/index.ts
@@ -1,1 +1,2 @@
 export * from './imports-config';
+export * from './settings-migration';